requirements.txt
python-a2s
disnake>=2.9.1
beautifulsoup4>=4.12.3
requests>=2.31.0
plotly
kaleido
